+++ database/migrations/2025_03_08_060141_create_implant_models_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('implant_model_sn');
             $table->integer('implant_model_qty')->default(1);
-            $table->double('implant_model_itemPrice')->default(0.00);
+            $table->decimal('implant_model_itemPrice', 10, 2)->default(0.00);
             $table->foreignId('implant_id')->constrained('implants');
             $table->foreignId('model_id')->constrained('abbott_models');
             $table->foreignId('stock_location_id')->constrained('stock_locations');
